separate fillTemplate into Utils
